requirements.txt
asgiref==3.5.2
Django==4.1.4
django-compat==1.0.15
django4-background-tasks==1.2.7
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
